requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.16
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
elasticsearch==7.15.2
idna==3.4
importlib-metadata==4.8.3
kombu==5.1.0
mutagen==1.45.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.33
py3Fdfs==2.2.0
pycparser==2.21
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.9
python-alipay-sdk==3.0.4
pytz==2022.6
redis==4.3.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.6.0
